+++ CHANGELOG.md
@@ -12,6 +12,9 @@ hi
 - Slightly decreased the spawning of Sea Anemone and Sea Whips when using Bone Meal underwater.
 - Sea Anemones can no longer be Bone Mealed directly, instead only being renewable by using Bone Meal underwater.
 - Slightly decreased the average amount of Pollen particles that spawn.
+- Deep Dark biome music will now play in Ancient City structures.
+  - This change was made to help with edge-cases, such as when an Ancient City partially generates into another biome.
+  - This change can be disabled in the config.
 - Music will now have a subtle, slowly morphing pitch shift while in the Dying Forest, Dying Mixed Forest, Snowy Dying Forest, and Snowy Dying Mixed Forest biomes.
 - `BetaBeachConditionSource` and `SnowUnderMountainConditionSource` no longer take any parameters and are now a single static instance.
 - Fixed the item tint of Willow Leaves on 1.21.4+.